backend/requirements.txt
amqp==5.3.1
bidict==0.23.1
billiard==4.2.4
blinker==1.9.0
boto3==1.42.73
botocore==1.42.73
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.6
decorator==5.2.1
dnspython==2.8.0
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-SocketIO==5.6.1
future==1.0.0
geocoder==1.38.1
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.3.2
gunicorn==25.1.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
kombu==5.6.2
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
prompt_toolkit==3.0.52
pycparser==3.0
pycryptodome==3.23.0
Pygments==2.20.0
PyJWT==2.12.1
pymongo==4.16.0
PyPDF2==3.0.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-http-client==3.3.7
python-socketio==5.16.1
ratelim==0.1.6
razorpay==2.0.1
redis==6.4.0
requests==2.32.5
s3transfer==0.16.0
sendgrid==6.12.5
simple-websocket==1.1.0
six==1.17.0
stripe==15.0.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
Werkzeug==3.1.6
wsproto==1.3.2
zope.event==6.1
zope.interface==8.2
